% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_analyte_summary}
\alias{get_analyte_summary}
\title{Get Analyte Summary}
\usage{
get_analyte_summary(psid)
}
\arguments{
\item{psid}{water system id}
}
\description{
For a given water system ID (psid), returns all the analytes
with their date range and number of records
}
\examples{
get_analyte_summary("0105009")
}
