% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Query SDWIS by PSID}
\usage{
get_data(psid, start_date = NULL, end_date = NULL, analyte = NULL)
}
\arguments{
\item{psid}{The public water system ID to be queried.}

\item{start_date}{The starting date for queried results.}

\item{end_date}{The end date for queried results.}

\item{analyte}{A single analyte to return.}
}
\value{
A data.table of analyte results for the input PSID and date range.
}
\description{
Query the SDWIS database by a public water system ID (PSID), start and end date, and analyte.
}
\examples{
d <- get_data("103039")
head(d)

}
