% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_storet_id}
\alias{get_storet_id}
\title{Get Storet ID}
\usage{
get_storet_id(analyte)
}
\arguments{
\item{analyte}{common name of chemical}
}
\description{
Helps users obtain the appropriate storet ID for a given analyte
}
\examples{
get_storet_id("c")
get_storet_id("Color")
get_storet_id("COLOR")
}
