% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_water_system}
\alias{get_water_system}
\title{Get Water System}
\usage{
get_water_system(county = NULL)
}
\arguments{
\item{county}{county}
}
\description{
Enable user to discover water system ID when supplied a county of interest
}
\examples{
get_water_system(county = "ALAMEDA")
}
