% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_psid.R
\name{query_psid}
\alias{query_psid}
\title{Query SDWIS by PSID}
\usage{
query_psid(psid, start_date = NULL, end_date = NULL, analytes = NULL)
}
\arguments{
\item{psid}{The public water system ID to be queried.}

\item{start_date}{The starting date for queried results.}

\item{end_date}{The end date for queried results.}

\item{analytes}{A vector of analytes to return.}
}
\value{
A tibble of analyte results for the input PSID and given dates.
}
\description{
Query the SDWIS database by a public water system ID (PSID), start and end date, and analyte.
}
\examples{
d <- query_psid("103039")
head(d)

query_psid(
  psid    = "110001",
  analyte = "NITRATE + NITRITE (AS N)"
) \%>\%
ggplot(aes(SAMP_DATE, FINDING)) +
  geom_point() +
  geom_smooth()
}
